requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
branca==0.8.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
crispy-bootstrap4==2024.1
diff-match-patch==20230430
distlib==0.3.8
Django==4.2
django-bootstrap4==24.4
django-crispy-forms==2.3
django-froala-editor==4.3.0
django-import-export==4.1.1
django-leaflet==0.31.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
filelock==3.15.1
folium==0.17.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
mysqlclient==2.2.4
numpy==2.1.2
pillow==10.4.0
platformdirs==4.2.2
reportlab==4.2.2
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
tablib==3.5.0
tzdata==2024.1
urllib3==2.2.3
virtualenv==20.26.2
xyzservices==2024.9.0
